getCowshedAngle

+++ src/d/actor/d_a_cow.cpp
@@ -1133,7 +1133,7 @@ int daCow_c::checkOutOfGate(cXyz pos) {
 
 /* 8065B8A8-8065B8D8 0033C8 0030+00 3/3 0/0 0/0 .text            getCowshedAngle__7daCow_cFv */
 s16 daCow_c::getCowshedAngle() {
-    return cLib_targetAngleY(&current.pos, &pen_pos);
+    return (s16)cLib_targetAngleY(&current.pos, &pen_pos);
 }
 
 /* 8065B8D8-8065BA30 0033F8 0158+00 1/1 0/0 0/0 .text            getCowshedDist__7daCow_cFv */
